requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
twilio==8.10.0
pyautogen==0.8.5
requests==2.31.0
python-multipart==0.0.6 